Add AEM Assets Support (#525)
* add AEM Assets script/plugin for image slots

* add conditional url generation for non-dropins

* move type declaration to .d.ts file

* fix crop settings example jsdoc

* ACAP-828 Add AEM Assets support in product-list-page (#389)

* refactor: rebase and remove async from config queries

* ACAP-1002 Add AEM Assets support in product-recommendations (#394)

* test: streamline and fix cypress tests

* fix: matching regular expressions and resolved protocol

* refactor: screen width in srcset image data

* ACAP-829: Enable AEM Assets in PDP (#420)

* chore: update pdp to 1.2.0-beta1

* chore: bump storefront sdk to v1.2.0

* feat: add support for AEM Assets images in PDP

* chore: bump pdp version to 1.2.0-beta2

* chore: pin sdk to 1.2.0

* test: add pdp aem assets tests

* refactor: try generate assets url passthrough for non aem assets + swatch image slot

* refactor: complete jsdoc

* refactor: streamline image slot config

* refactor: remove wrapper from slot config maker

* fix: update cypress tests to work with adobe store products

* refactor: allow more truthy values in aem-assets-enabled config

* fix: potential undefined config value invoking function

* refactor: true passthrough on try methods and more robust config check

* ACAP-942: Enable AEM Assets in Account (#446)

* ACAP-1027: Order drop in compatibility (#454)

* add asset image rendering to order returns

* add asset image rendering to returns list

* add asset image rendering to orders list and shipping status

* add test

* update storefront-order dropin in

* refactor test and duplicated code

* refactor imageslotconfig func

* ACAP-1004: Enable AEM Assets in Cart (#428)

* ACAP-829: Enable AEM Assets in PDP (#420)

* chore: update pdp to 1.2.0-beta1

* chore: bump storefront sdk to v1.2.0

* feat: add support for AEM Assets images in PDP

* chore: bump pdp version to 1.2.0-beta2

* chore: pin sdk to 1.2.0

* test: add pdp aem assets tests

* refactor: try generate assets url passthrough for non aem assets + swatch image slot

* refactor: complete jsdoc

* refactor: streamline image slot config

* refactor: remove wrapper from slot config maker

* feat: add cart thumbnail slot

* feat: add mini cart thumbnail slot

* feat: add gift options swatch image slot

* test: add cart aem assets tests

* test: skip aem assets tests

* refactor: fix linting issues

* refactor: use destructured item instead of ctx

* chore: sync with latest ACAP-1017

* fix: swatch image slots

* fix: sync assets script and lint issues

* fix: newline end assets scripts

* refactor: consistent image slots

* refactor: extract swatch image slot to reusable function

* ACAP-1032: Add support for aem images in checkout (#519)

* Add slots for aem assets rendering

* Added test for checkout

* fix: getConfigValue import path

* refactor: support AEM Assets in custom product list

* test: fix aem assets tests

* test: fix aem assets tests and PR workflows

* chore: remove PREX support in favor of drop-in version

* chore: add product-recommendations back clean from integration

* test: complete cypress tests

* chore: change env secret names

* chore: temp remove json loading secrets

* chore: comment json.parse

* fix: removed lines from product recommendations block

* chore: update @dropins/tools to version 1.3.0-alpha16 and refactor as… (#529)

* chore: update @dropins/tools to version 1.3.0-alpha16 and refactor asset rendering imports

- Updated the version of @dropins/tools in package.json and package-lock.json to 1.3.0-alpha16.
- Refactored multiple components to import `tryRenderAemAssetsImage` from the updated @dropins/tools library instead of local scripts.
- Removed obsolete assets.d.ts and assets.js files, consolidating AEM asset handling into the new structure.

* chore: update @dropins/tools to version 1.3.0-beta3 and remove obsolete TypeScript definitions

- Updated the version of @dropins/tools in package.json and package-lock.json to 1.3.0-beta3.
- Removed obsolete TypeScript definition files related to fetch-graphql and recaptcha, streamlining the codebase.

---------

+++ cypress/src/support/deleteCustomer.js
@@ -25,5 +25,9 @@ Cypress.Commands.add('deleteCustomer', deleteCustomer);
 
 // Always delete customer after every test.
 afterEach(() => {
+  if (Cypress.env("isAemAssetsSuite")) {
+    return;
+  }
+
   cy.deleteCustomer();
 });
